python/requirements.txt
-i https://pypi.org/simple
aioredis==1.3.1
async-timeout==3.0.1
certifi==2020.11.8
chardet==3.0.4
fastapi==0.62.0
h11==0.8.1
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
loguru==0.5.3
pydantic==1.7.3
python-dotenv==0.15.0
requests-async==0.6.2
requests==2.25.0
rfc3986==1.4.0
starlette==0.13.6
tenacity==6.3.1
urllib3==1.26.2
